// front/src/features/rrhh/types.ts
export type RrhhUserDocument = {
  id: number;
  userId: number;
  category: string | null;
  title: string | null;
  description: string | null;
  fechaVencimiento: string | null;
  mime: string | null;
  size: number | null;
  originalName: string | null;
  uploadedBy: number | null;
  createdAt: string | null;
  downloadUrl: string | null;
};